WIP migrating literary and lyrical corpora into lib folder.
